refactor: remove instant mask due to formatter implementation
