--- src/utils/masks/currency/instant/index.ts
@@ -1,8 +1,0 @@
-export const maskCurrencyInstant = (value: number) => {
-  if (isNaN(value)) {
-    return '';
-  }
-
-  const options = { currency: 'BRL', style: 'currency' };
-  return new Intl.NumberFormat('pt-BR', options).format(value);
-};
